refactor(gateway): move shared URL/probe utilities to worker/util.rs
classify.rs (parent-level) was importing http_base_url, try_http_reachable,
and try_grpc_reachable from worker::local — wrong dependency direction.
A parent-level orchestration step should not reach into a branch-specific
submodule for shared utilities.

Move strip_protocol, http_base_url, grpc_base_url, try_http_reachable,
do_grpc_health_check, and try_grpc_reachable to worker/util.rs. Update
all import paths in classify.rs, detect_connection.rs, detect_backend.rs,
and discover_metadata.rs.

Also fix is_models_endpoint_reachable: use !status.is_server_error()
instead of enumerating 2xx/401/403. The old filter missed 429 (rate
limited), which would misclassify a rate-limited cloud API as Local.
