+++ model_gateway/src/core/steps/worker/local/detect_connection.rs
@@ -4,91 +4,17 @@
 //! This step only answers "HTTP or gRPC?" — backend runtime detection
 //! (sglang vs vllm vs trtllm) is handled by the separate DetectBackendStep.
 
-use std::time::Duration;
-
 use async_trait::async_trait;
-use reqwest::Client;
 use tracing::debug;
 use wfaas::{StepExecutor, StepId, StepResult, WorkflowContext, WorkflowError, WorkflowResult};
 
-use super::strip_protocol;
-use crate::{
-    core::{
-        steps::workflow_data::{WorkerKind, WorkerWorkflowData},
-        ConnectionMode,
+use crate::core::{
+    steps::{
+        worker::util::{try_grpc_reachable, try_http_reachable},
+        workflow_data::{WorkerKind, WorkerWorkflowData},
     },
-    routers::grpc::client::GrpcClient,
+    ConnectionMode,
 };
-
-/// Try HTTP health check.
-pub(crate) async fn try_http_reachable(
-    url: &str,
-    timeout_secs: u64,
-    client: &Client,
-) -> Result<(), String> {
-    let is_https = url.starts_with("https://");
-    let protocol = if is_https { "https" } else { "http" };
-    let clean_url = strip_protocol(url);
-    let health_url = format!("{protocol}://{clean_url}/health");
-
-    client
-        .get(&health_url)
-        .timeout(Duration::from_secs(timeout_secs))
-        .send()
-        .await
-        .and_then(reqwest::Response::error_for_status)
-        .map_err(|e| format!("Health check failed: {e}"))?;
-
-    Ok(())
-}
-
-/// Perform a single gRPC health check with a specific runtime type.
-///
-/// Shared with `detect_backend` which uses this for runtime identification.
-pub(super) async fn do_grpc_health_check(
-    grpc_url: &str,
-    timeout_secs: u64,
-    runtime_type: &str,
-) -> Result<(), String> {
-    let connect_future = GrpcClient::connect(grpc_url, runtime_type);
-    let client = tokio::time::timeout(Duration::from_secs(timeout_secs), connect_future)
-        .await
-        .map_err(|_| "gRPC connection timeout".to_string())?
-        .map_err(|e| format!("gRPC connection failed: {e}"))?;
-
-    let health_future = client.health_check();
-    tokio::time::timeout(Duration::from_secs(timeout_secs), health_future)
-        .await
-        .map_err(|_| "gRPC health check timeout".to_string())?
-        .map_err(|e| format!("gRPC health check failed: {e}"))?;
-
-    Ok(())
-}
-
-/// Check if gRPC is reachable by trying all known runtime types in parallel.
-///
-/// We don't care which runtime it is here — that's detect_backend's job.
-/// We just need to know: does this endpoint speak gRPC at all?
-pub(crate) async fn try_grpc_reachable(url: &str, timeout_secs: u64) -> Result<(), String> {
-    let grpc_url = if url.starts_with("grpc://") {
-        url.to_string()
-    } else {
-        format!("grpc://{}", strip_protocol(url))
-    };
-
-    let (sglang, vllm, trtllm) = tokio::join!(
-        do_grpc_health_check(&grpc_url, timeout_secs, "sglang"),
-        do_grpc_health_check(&grpc_url, timeout_secs, "vllm"),
-        do_grpc_health_check(&grpc_url, timeout_secs, "trtllm"),
-    );
-
-    match (sglang, vllm, trtllm) {
-        (Ok(()), _, _) | (_, Ok(()), _) | (_, _, Ok(())) => Ok(()),
-        (Err(e1), Err(e2), Err(e3)) => Err(format!(
-            "gRPC not reachable (tried sglang, vllm, trtllm): sglang={e1}, vllm={e2}, trtllm={e3}",
-        )),
-    }
-}
 
 /// Step 1: Detect connection mode (HTTP vs gRPC).
 ///
